Conflicts:
	config/warble.rb

+++ db/migrate/20101202094602_add_origin_to_collections.rb
@@ -20,6 +20,6 @@ class AddOriginToCollections < ActiveRecord::Migration
   end
 
   def self.down
-    remove_column :collections, :origin
+    remove_column :collections, :origin if column_exists? :collections, :origin
   end
 end
